backend/requirements.txt
Django>=3.0,<4.0
djangorestframework
gunicorn
psycopg2-binary
